requirements.txt
pyyaml==3.11
pytest==3.1.2
tox==2.3.1
cryptography==1.7
cookiecutter>=1.4.0
pytest-cookies==0.2.0
watchdog==0.8.3
alabaster==0.7.8
fabric3